% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_3PGhydro.R
\name{run_3PGhydro}
\alias{run_3PGhydro}
\title{Run 3PG-Hydro}
\usage{
run_3PGhydro(
  climate,
  p,
  lat,
  StartDate,
  StandAgei,
  EndAge,
  WFi,
  WRi,
  WSi,
  StemNoi,
  CO2Concentration,
  FR,
  HeightEquation,
  SVEquation,
  SoilClass,
  EffectiveRootZoneDepth,
  DeepRootZoneDepth,
  RocksER,
  RocksDR,
  thinAges,
  thinVals,
  thinWF,
  thinWR,
  thinWS
)
}
\arguments{
\item{climate}{Climate data as .csv file, mandatory column names ("date" ("dd/mm/yyyy"),"Tav","Tmax","Tmin","Rain","SolarRad")}

\item{p}{3PG tree species parameter .csv file}

\item{lat}{site latitude UTM}

\item{StartDate}{starting date, format: "dd/mm/yyyy"}

\item{StandAgei}{starting stand age in years}

\item{EndAge}{end stand age in years}

\item{WFi}{inital weight of foliage in tons}

\item{WRi}{inital weight of roots in tons}

\item{WSi}{inital weight of stems in tons}

\item{StemNoi}{initial stem number}

\item{CO2Concentration}{select which CO2 concentration equation should be used: "Historical", "RCP2.6", "RCP4.5", "RCP8.5"}

\item{FR}{fertility rating of site (0-1)}

\item{HeightEquation}{choose the height equation (1-2), (1: 3PG Original Equation, 2: Michajlow-Schumacher from Forrester et al. (2021))}

\item{SVEquation}{choose the equation used for stand volume calculation (1-3) (1: V = aV * H^VH * D^VB, Forrester et al. (2021), 2: 3PG original function with stem mass & density, 3: simple volume calculation V = FormFactor * CylinderVolume)}

\item{SoilClass}{soil classes, type number 1 - 4 (1: sand, 2: sandy loam, 3: clay loam, 4: clay)}

\item{EffectiveRootZoneDepth}{depth of effective root zone in meter}

\item{DeepRootZoneDepth}{depth of deep root zone in meter}

\item{RocksER}{share of rocks/skeleton in effective root zone (0-1) 0:none, 1:all}

\item{RocksDR}{share of rocks/skeleton in deep root zone (0-1) 0:none, 1:all}

\item{thinAges}{ages at which thinning should be applied, in case of no thinning: NULL}

\item{thinVals}{number of stems left after thinning, in case of no thinning: NULL}

\item{thinWF}{thinning regime foliage below to above (0.5 - 1.5), in case of no thinning: NULL}

\item{thinWR}{thinning regime foliage below to above (0.5 - 1.5), in case of no thinning: NULL}

\item{thinWS}{thinning regime foliage below to above (0.5 - 1.5), in case of no thinning: NULL}
}
\value{
output file in daily time steps: Date, StandAge, StemNo, WF, WR, WS, DBH, Height, StandVol, volWCer, volWCdr, NPP, NEE, LAI, Evapotranspiration, AvStemMass, Basal Area, Self Thinning, WSext, StandVol_loss, VolProduction_tot, Deep Percolation, Run Off
}
\description{
3PG-Hydro is an update of the original 3PG Forest Growth model by Landsberg and Waring (1997) (DOI:10.1016/S0378-1127(97)00026-1). 3PG-Hydro calculates on a daily timestep, includes a soil-water-model as well as a snow routine. Further information in the publication on 3PG-Hydro by Yousefpour and Djahangard (2021) (DOI:10.3390/f12121729). 3PG-Hydro is available as an R-package, coding done by Anja Nölte & Marc Djahangard.
}
\examples{
climate <- read.csv("climate.csv")
p <- read.csv("3PG_Parameter.csv") 
p <- p$species1 #select species parameters column
lat <- 40.5
StartDate <- "01/01/1980"
StandAgei <- 40
EndAge <- 65
StemNoi <- 300
WSi <- 150
WFi <- 10
WRi <- 20
CO2Concentration <- "Historical"
FR <- 0.7
HeightEquation <- 1
SVEquation <- 2
SoilClass <- 2
EffectiveRootZoneDepth <- 1
DeepRootZoneDepth <- 5
RocksER <- 0.2 
RocksDR <- 0.4 
thinAges <- c(50,55,60)
thinVals <- c(250,200,150) 
thinWF <- rep(0.8,length(thinAges)) 
thinWR <- rep(0.8,length(thinAges)) 
thinWS <- rep(0.8,length(thinAges))  
out <- run_3PGhydro(climate,p,lat,StartDate,StandAgei,EndAge,WFi,WRi,WSi,StemNoi,CO2Concentration,FR,SoilClass,EffectiveRootZoneDepth,DeepRootZoneDepth,RocksER,RocksDR,thinAges,thinVals,thinWF,thinWR,thinWS)
}
